% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{buildCITE}
\alias{buildCITE}
\title{function to build Seurat Object with CITE seq files}
\usage{
buildCITE(rna, adt)
}
\arguments{
\item{rna}{gz file name for RNA generated using CITE-seq}

\item{adt}{gz file name for ADT generated using CITE-seq}
}
\value{
seurat object with RNA and ADT assays
}
\description{
function to build Seurat Object with CITE seq files
}
