% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{checkFeatures}
\alias{checkFeatures}
\title{Function to return list of features found in assay for Seurat object}
\usage{
checkFeatures(obj, features, assay)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{features to check for}

\item{assay}{assay in Seurat object to check for features}
}
\value{
filtered list of features with only features found in object's assay
}
\description{
Function to return list of features found in assay for Seurat object
}
