% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTFs.R
\name{estProtein}
\alias{estProtein}
\title{Estimate target protein expression for TF}
\usage{
estProtein(x, obj, db.filt, tf, ident)
}
\arguments{
\item{x}{vector with protein-gene scaling factor per cluster}

\item{tf}{transcription factor to estimate}
}
\description{
Estimate target protein expression for TF
}
