% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTFs.R
\name{getScale}
\alias{getScale}
\title{Estimate protein expression by calculating scaling factor per cluster}
\usage{
getScale(obj, ident, sampleSize = 100)
}
\description{
Estimate protein expression by calculating scaling factor per cluster
}
