% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getSIG}
\alias{getSIG}
\title{Function to assist with voltf() and plotProteins()}
\usage{
getSIG(df, logT, pT)
}
\arguments{
\item{df}{dataframe output from getTargets()}

\item{pT}{FDR threshold for significance}

\item{log2FC}{threshold for significance}
}
\value{
a dataframe with columns added SIG (UP, DOWN, NS) and show (label names)
}
\description{
Function to assist with voltf() and plotProteins()
}
