% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTFs.R
\name{getTargets}
\alias{getTargets}
\title{Function to retrieve Log2FC values for TF targets in ATAC based on input topTFs for cluster/groups in RNA}
\usage{
getTargets(tfs, markers, db)
}
\arguments{
\item{tfs}{output from getTFs, list of top TFs per cluster/group}

\item{markers}{ATAC top markers per cluster/group, output from ArchR::getMarkers() function}

\item{db}{TF-target database with names(db) as TFs}
}
\description{
Function to retrieve Log2FC values for TF targets in ATAC based on input topTFs for cluster/groups in RNA
}
