% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTFs.R
\name{getTFs}
\alias{getTFs}
\title{Function to retrieve top upregulated TFs for RNA clusters/groups}
\usage{
getTFs(obj, markers, tfs, num = 20)
}
\arguments{
\item{obj}{Seurat object with RNA data, active ident should be what markers are comparing (ex: seurat_clusters)}

\item{markers}{differentially expressed genes by cluster or group}

\item{num}{number of TFs to export per cluster/group, default is 20}

\item{db}{TF-target database with names(db) as TFs}
}
\description{
Function to retrieve top upregulated TFs for RNA clusters/groups
}
