% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEn}
\alias{plotEn}
\title{Plots enrichment of TF gene set}
\usage{
plotEn(markers, db.filt, path)
}
\arguments{
\item{markers}{atac DE markers}

\item{db.filt}{TF-target database}

\item{path}{TF of interest}
}
\value{
enrichment plot generated using fgsea
}
\description{
Plots enrichment of TF gene set
}
