% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotProteins}
\alias{plotProteins}
\title{Function to plot protein levels of significant TF targets in different clusters/groups}
\usage{
plotProteins(targets, type, obj, filename, ident, logT = 0.5, pT = 0.01)
}
\description{
Function to plot protein levels of significant TF targets in different clusters/groups
}
