% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tfBar}
\alias{tfBar}
\title{function to generate boxplots of estimated protein expression of TF targets}
\usage{
tfBar(est, tf)
}
\arguments{
\item{est}{estimated protein expressin output from estProteins() function}

\item{tf}{transcription factor of interest}
}
\value{
boxplot with expression values and signficance from ANOVA or t.test
}
\description{
function to generate boxplots of estimated protein expression of TF targets
}
