% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tfMap}
\alias{tfMap}
\title{Function to plot heatmap for top 10 TFs found differentially expressed (RNA) per cluster/grouping}
\usage{
tfMap(obj, markers, tfs, type = "both")
}
\arguments{
\item{obj}{Seurat object with RNA data, active ident should be what markers are comparing (ex: seurat_clusters)}

\item{markers}{differentially expressed genes by cluster or group}

\item{type}{type of markers to plot; options are upregulated (up), downregulated (down), or default both}

\item{db}{TF database with names(db) as TFs}
}
\value{
ggplot object with Heatmap
}
\description{
Function to plot heatmap for top 10 TFs found differentially expressed (RNA) per cluster/grouping
}
