% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tfSimilarity}
\alias{tfSimilarity}
\title{Function to plot a matrix with Jaccard Similarity values based on how similar up-regulated TFs are between clusters/groups}
\usage{
tfSimilarity(obj, markers, db, plot = T)
}
\arguments{
\item{obj}{Seurat object with RNA data, active ident should be what markers are comparing (ex: seurat_clusters)}

\item{markers}{differentially expressed genes by cluster or group}

\item{db}{TF database with names(db) as TFs}

\item{plot}{logical to indicate whether to return plot (TRUE) or return similarity matrix (FALSE)}
}
\value{
either ggplot object or Jaccard Similarity matrix
}
\description{
Function to plot a matrix with Jaccard Similarity values based on how similar up-regulated TFs are between clusters/groups
}
