% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{voltf}
\alias{voltf}
\title{Function to make volcano plot for each top upregulated TF in a group/cluster}
\usage{
voltf(targets, db, tfs, ident, filename, logT = 0.5, pT = 0.01)
}
\arguments{
\item{targets}{dataframe output from getTargets()}

\item{ident}{ident to plot TF-targets for}

\item{filename}{filename for saving pdf with all plots}

\item{logT}{log2FC threshold of significance for volcano plot (default = 0.5)}

\item{pT}{FDR threshold of significance for volcano plot (default = 0.01)}
}
\value{
saves pdf with filename
}
\description{
Function to make volcano plot for each top upregulated TF in a group/cluster
}
